% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Calculate_Cluster_Centroids}
\alias{Calculate_Cluster_Centroids}
\title{Generic function to compute distances, centroids and statistics after statistics.}
\usage{
Calculate_Cluster_Centroids(X)
}
\arguments{
\item{X}{an MCXpress object with MCA object (after MCA step)}
}
\value{
MCXpress object with MCA object and Clustering Object (after MCA step)
}
\description{
Generic function to compute distances, centroids and statistics after statistics.
}
