% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_hclust}
\alias{cluster_hclust}
\title{Hierarchical Clustering on MCA}
\usage{
cluster_hclust(X, method = "average", k = NULL, h = NULL)
}
\arguments{
\item{X}{MCXpress object containing a MCXmca object}

\item{method}{hclust method ("average", "ward", etc..)}

\item{k}{integer indicating the number of cluster to obtain}

\item{h}{numeric indicating the heights where the tree should be cut}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{labels}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{gene_cluster_distances}{GSEA Parameter used}
\item{closest_cluster}{Indicate the closest cluster for a gene}
\item{coord_centroids}{GSEA Parameter used}
\item{plot1}{Clustering Visualisation in the MCA cell space}
\item{plot2}{Visualisation of the centroids in the MCA gene space}
}
\description{
Hierarchical Clustering on MCA
}
