% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_igraph}
\alias{cluster_igraph}
\title{K-Means Clustering on MCA}
\usage{
cluster_igraph(X, dim = 2, max_distance = Inf, ratio = 1, k = NULL,
  mutual = TRUE, methods = "louvain")
}
\arguments{
\item{X}{MCXpress object containing MCA object}

\item{dim}{number of component to retain for distance calculation.}

\item{max_distance}{Distance threshold for edges conservation}

\item{ratio}{Percentages of the closest cells to be linked to}

\item{k}{number of nearest neighbour}

\item{mutual}{nearest neighbour mutual or not}

\item{methods}{clustering methods for graph}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{labels}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{gene_cluster_distances}{GSEA Parameter used}
\item{closest_cluster}{Indicate the closest cluster for a gene}
\item{coord_centroids}{GSEA Parameter used}
\item{plot1}{Clustering Visualisation in the MCA cell space}
\item{plot2}{Visualisation of the centroids in the MCA gene space}
}
\description{
K-Means Clustering on MCA
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553)
MCX64553 <- cluster_igraph(MCX64553,dim=3, k=3, mutual=TRUE,methods="fast_greedy")
}
