% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_k_medoids}
\alias{cluster_k_medoids}
\title{K-Medoids Clusetering on the MCA Analysis}
\usage{
cluster_k_medoids(X, k = 2)
}
\arguments{
\item{X}{MCXpress object containing a MCXmca object}

\item{k}{integer indicating the number of cluster to obtain}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{labels}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{gene_cluster_distances}{GSEA Parameter used}
\item{closest_cluster}{Indicate the closest cluster for a gene}
\item{coord_centroids}{GSEA Parameter used}
\item{plot1}{Clustering Visualisation in the MCA cell space}
\item{plot2}{Visualisation of the centroids in the MCA gene space}
}
\description{
K-Medoids Clusetering on the MCA Analysis
}
