% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{K-Means Clustering on MCA}
\usage{
cluster_kmeans(X, k = 2, maxIter = 10, nstart = 50)
}
\arguments{
\item{X}{MCXpress object containing a MCXmca object}

\item{k}{Number of cluster to compute}

\item{maxIter}{maximum number of iteration}

\item{nstart}{number of random sets of initial centers}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{labels}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{gene_cluster_distances}{GSEA Parameter used}
\item{closest_cluster}{Indicate the closest cluster for a gene}
\item{coord_centroids}{GSEA Parameter used}
\item{plot1}{Clustering Visualisation in the MCA cell space}
\item{plot2}{Visualisation of the centroids in the MCA gene space}
}
\description{
K-Means Clustering on MCA
}
