% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_supervised}
\alias{cluster_supervised}
\title{Supervised Clustering on MCA}
\usage{
cluster_supervised(X, Y)
}
\arguments{
\item{X}{MCXpress object containing Dim_Red object}

\item{Y}{A vector with Sample names as name and the coresponding cluster as value.}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{Cluster_Quali}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{Gene_Cluster_Distance}{GSEA Parameter used}
\item{closest_Cluster}{Indicate the closest cluster for a gene}
\item{Coord_Centroids}{GSEA Parameter used}
\item{graph1}{Clustering Visualisation in the MCA cell space}
\item{graph2}{Visualisation of the centroids in the MCA gene space}
}
\description{
Supervised Clustering on MCA
}
