% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA.R
\name{Dimension_reduction_MCA}
\alias{Dimension_reduction_MCA}
\title{Performs Multiple Corespondence Analysis on a object of class MCXpress containing a Disjunctive Matrix.}
\usage{
Dimension_reduction_MCA(X, Dim = 5)
}
\arguments{
\item{X}{A MCXpress class object containing a Disjunctive Matrix.}

\item{Dim}{number of Axis to be kept for the variables and individual coordinates and used for the distance calculation.}
}
\description{
Performs Multiple Corespondence Analysis on a object of class MCXpress containing a Disjunctive Matrix.
}

