% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection_Methods.R
\name{filter_outlier}
\alias{filter_outlier}
\title{Filter genes that are expressed poorly in a few cells}
\usage{
filter_outlier(X, percentage = 0.05, threshold = 1)
}
\arguments{
\item{X}{MCXpress Object with an Expression Matrix}

\item{percentage}{numeric between 0 and 1 indicating the percentages of cells that must express the genes.}

\item{threshold}{gene expression lower threshold.}
}
\value{
A filtered Expression Matrix in the MCXpressObject.
}
\description{
Filter genes that are expressed poorly in a few cells
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
}
