% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imports.R
\docType{data}
\name{GSE64553}
\alias{GSE64553}
\title{Expression Matrix from GEO ID: GSE64553}
\format{A matrix with 47966 genes (row) and 35 individuals (col)}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE64553}
}
\usage{
GSE64553
}
\description{
The dataset contains only the value for HFF samples
and one outlier has been removed.
}
\keyword{datasets}

