% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{GSEA_Heatmap_Cluster}
\alias{GSEA_Heatmap_Cluster}
\title{Title}
\usage{
GSEA_Heatmap_Cluster(X, pval = 0.05, es = 0, nes = -20,
  color = cm.colors(100), title = "", rmna = T, nPath = 5,
  metrics = "NES", plotly = F, margin = c(0, 0, 0, 0), cexCol = 1,
  cexRow = 1, row_color = rainbow)
}
\arguments{
\item{plotly}{}
}
\description{
Title
}
