% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{GSEA_Heatmap_SC_Cluster}
\alias{GSEA_Heatmap_SC_Cluster}
\title{Heatmap of GSEA at Single Cell Level}
\usage{
GSEA_Heatmap_SC_Cluster(X, pval = 0.05, es = 0, nes = -20,
  color = cm.colors(100), title = "", rmna = T, metrics = "NES",
  plotly = F)
}
\arguments{
\item{pval}{Numeric indicating the threshold pvalue}

\item{es}{Numeric indicating enrichment score threshold}

\item{nes}{Numeric normalised enrichment score threshold}

\item{color}{color palette to use for the heatmap}

\item{title}{Plot Title}

\item{rmna}{Logical indicating if any empty column should be removed.}

\item{metrics}{NES or ES}

\item{plotly}{Logical Indicating if the static version or the interactive version should be plotted.}

\item{x}{An MCXpress object after clustering and SC_GSEAparall step}
}
\value{
Heatmap with enrichment score of each cells ordered according to hierchical clustering inside each cluster.
}
\description{
Heatmap of GSEA at Single Cell Level
}
