% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Hallmark}
\alias{Hallmark}
\title{gmtfile from reactome pathway}
\format{A list of 50 Hallmark geneset with hg19 symbol.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
Hallmark
}
\description{
Contains all the geneset of from Broad Institute Hallmark.
}
\keyword{datasets}
