% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Heatmap_Cluster}
\alias{Heatmap_Cluster}
\title{Title}
\usage{
Heatmap_Cluster(X, n = 5, plotly = T)
}
\arguments{
\item{X}{MCXpress Object after Clustering}

\item{n}{Number of Genes to Display per Cluster}

\item{plotly}{Logical indicating if interactive html visualisation should be used}
}
\value{

}
\description{
Title
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553, Dim = 5)
MCX64553 <- cluster_kmeans(MCX64553, k=6)
MCX64553 \%>\% Heatmap_Cluster(n = 5, plotly = F)
}
