% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Heatmap_Cluster_SC}
\alias{Heatmap_Cluster_SC}
\title{Finding Top Expressed Genes for Each Cluster}
\usage{
Heatmap_Cluster_SC(x, n = 5, plotly = F)
}
\arguments{
\item{n}{Integer specifying the number of genes to select for each cluster, default set at 5.}

\item{plotly}{logical, if FALSE gives a static heatmap, if TRUE gives an interactive version with heatmaply.}

\item{X}{An MCXpress Object}
}
\value{
A Heatmap Visualisation
}
\description{
Search for each cluster the Top n closest genes in the MCA components and visualise it on a heatmap.
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553, Dim = 5)
MCX64553 <- cluster_kmeans(MCX64553, k=6)
MCX64553 \%>\% Heatmap_Cluster_SC(n = 5, plotly = F)
}
