% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Initialise_MCXpress}
\alias{Initialise_MCXpress}
\title{Initialisation of the MCXpress Object}
\usage{
Initialise_MCXpress(X, min_reads = NULL)
}
\arguments{
\item{X}{DESCRIPTION.}

\item{min_reads}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Create a MCXpressObject from an Expression Matrix
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
}
