% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA.R
\name{MCA}
\alias{MCA}
\title{Dimensional reduction using multiple corespondence analysis.}
\usage{
MCA(X, Dim = (X$ExpressionMatrix \%>\% dim \%>\% min) - 1)
}
\arguments{
\item{X}{A MCXpress object containing a Disjunctive Matrix produced by one of the Discretisation method.}

\item{Dim}{Number of Dimension to keep.}
}
\value{
A MCXpress object containing a MCXmca object.
\itemize{
  \item{'cells_principal'}{Stuff}
  \item{'parameter 2'}{Stuff}
 }
}
\description{
Apply fuzzy multiple correspondance analysis on the disjunctive matrix to reduce the number of dimension
and keep the axis with the most informative variance
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553, Dim = 5)
}
