% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Num_Axis_Eigen_Distance}
\alias{Num_Axis_Eigen_Distance}
\title{Finding Top Expressed Genes for Each Cluster}
\usage{
Num_Axis_Eigen_Distance(X)
}
\arguments{
\item{X}{An MCXpress Object}
}
\value{
Number of Axis to be retained
}
\description{
Search for each cluster the Top n closest genes in the MCA components and visualise it on a heatmap.
}
