% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plotlyEnrichment}
\alias{plotlyEnrichment}
\title{Interactive plot enrichment score of a pathway}
\usage{
plotlyEnrichment(pathway, stats, gseaParam = 0)
}
\arguments{
\item{pathway}{A single geneset}

\item{stats}{Ranking of the genes}

\item{gseaParam}{GSEA parameter value}
}
\value{
RETURN_DESCRIPTION
}
\description{
Plotly version of the plotEnrichment function of the fgsea package
}
