% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imports.R
\docType{data}
\name{reactome_gmtfile}
\alias{reactome_gmtfile}
\title{gmtfile from reactome pathway}
\format{A list of 1892 reactome pathway geneset with hg19 symbol.}
\source{
\url{http://www.reactome.org/pages/download-data/}
}
\usage{
reactome_gmtfile
}
\description{
Contains all the geneset of the reactome pathway for human.
}
\keyword{datasets}
