% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{SC_GSEAparall}
\alias{SC_GSEAparall}
\title{GSEA at single cell level}
\usage{
SC_GSEAparall(X, GMTfile, nperm = 1000, minSize = 15, maxSize = 500,
  nproc = 4, nbin = 1, naxis = 2, gseaParam = 0, dim = 2)
}
\arguments{
\item{X}{A MCXPress object containing a Dimred object}

\item{GMTfile}{named list of pathway with their representing gene.}

\item{nperm}{the number of permutation.}

\item{minSize}{threshold for the minimum number of genes that must be in the pathway.}

\item{maxSize}{threshold for the maximum number of genes that must be in the pathway.}

\item{nproc}{number of processor to use for the enrichment algorithm.}

\item{nbin}{an integer indicating which bin to use for the enrichment analysis.}

\item{naxis}{number of axis to perform the enrichment analysis.}

\item{gseaParam}{an integer to weight the enrichment analysis.}

\item{dim}{Number of Axis to retain for the calculation gene ranking for each cells}
}
\value{
Return a MCXpress object containing a GSEA object
\item{GSEA_Results}{List containing dataframes of the GSEA analysis for each Cluster}
\item{Ranking}{List containing vectors of gene ranking according to their distance in MCA for each Cluster}
\item{GSEA_Results_Axis}{List containing dataframes of the GSEA analysis for each Axis}
\item{Ranking_Axis}{List containing vectors of gene ranking according to their MCA cells coordinate and genes expression correlation for each Axis}
\item{GMTfile}{GMTfile used for GSEA}
\item{Pathway}{Name of the genesets contained in the gmtfile}
\item{gseaParam}{GSEA Parameter used}
}
\description{
Compute the enrichment score of every cell to a given pathway set using the gene ranking derived from MCA.
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553)
MCX64553 <- SC_GSEAparall(X = MCX64553, GMTfile = reactome_gmtfile, nperm = 10000, dim=3)
}
