% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection_Methods.R
\name{select_most_variable_genes}
\alias{select_most_variable_genes}
\title{Select most variable genes}
\usage{
select_most_variable_genes(X, ngenes = NULL)
}
\arguments{
\item{X}{A MCXpress class object}

\item{ngenes}{Number of genes to keep, if not specified will take an  optimal number.}
}
\value{
A MCXpress object with an updated Expression_Matrix
}
\description{
Select most variable genes
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
}
