% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{calculate_cluster_centroids}
\alias{calculate_cluster_centroids}
\title{Generic function to compute distances, centroids and statistics after statistics.}
\usage{
calculate_cluster_centroids(X, dim)
}
\arguments{
\item{X}{an MCXpress object with MCA object (after MCA step)}
}
\value{
MCXpress object with MCA object and Clustering Object (after MCA step)
}
\description{
Generic function to compute distances, centroids and statistics after statistics.
}
