% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_k_medoids}
\alias{cluster_k_medoids}
\title{K-Medoids Clusetering on the MCA Analysis}
\usage{
cluster_k_medoids(X, k = 2)
}
\arguments{
\item{X}{MCXpress object containing a MCXmca object}

\item{k}{integer indicating the number of cluster to obtain}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{Cluster_Quali}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{Gene_Cluster_Distance}{GSEA Parameter used}
\item{closest_Cluster}{Indicate the closest cluster for a gene}
\item{Coord_Centroids}{GSEA Parameter used}
\item{graph1}{Clustering Visualisation in the MCA cell space}
\item{graph2}{Visualisation of the centroids in the MCA gene space}
}
\description{
K-Medoids Clusetering on the MCA Analysis
}
