% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_supervised}
\alias{cluster_supervised}
\title{Supervised Clustering on MCA}
\usage{
cluster_supervised(X, Y)
}
\arguments{
\item{X}{MCXpress object containing MCA object}

\item{Y}{A vector with Sample names as name and the coresponding cluster as value.}
}
\value{
MCXpress object containing a MCXmca and MCXcluster object
\item{labels}{Clustering results}
\item{nClusters}{Number of Cluster}
\item{gene_cluster_distances}{GSEA Parameter used}
\item{closest_cluster}{Indicate the closest cluster for a gene}
\item{coord_centroids}{GSEA Parameter used}
\item{plot1}{Clustering Visualisation in the MCA cell space}
\item{plot2}{Visualisation of the centroids in the MCA gene space}
}
\description{
Supervised Clustering on MCA
}
\examples{
MCX64553 <- Initialise_MCXpress(GSE64553)
MCX64553 <- filter_outlier(MCX64553, percentage = 0.05, threshold = 3)
MCX64553 <- discretisation_01(MCX64553, scaled=FALSE)
MCX64553 <- MCA(MCX64553)
your_cluster <- gsub("\\\\d$|HFF_",replacement = "", x=colnames(MCX64553$ExpressionMatrix))
names(your_cluster) <- colnames(MCX64553$ExpressionMatrix)
MCX64553 <-  cluster_supervised(MCX64553, your_cluster)
}
