% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{GSEA_Heatmap_Cluster}
\alias{GSEA_Heatmap_Cluster}
\title{Heatmap of Cluster Enrichement}
\usage{
GSEA_Heatmap_Cluster(X, pval = 0.05, es = 0, nes = -20,
  color = cm.colors(100), title = "", rmna = T, metrics = "NES",
  plotly = F, margin = c(0, 0, 0, 0))
}
\arguments{
\item{X}{MCX object with a cluster object inside}

\item{pval}{adjusted pvalue threshold}

\item{es}{enrichment score threshold}

\item{nes}{normalised enrichment score threshold}

\item{color}{color palette to use for heatmap}

\item{title}{title of the Heatmap}

\item{rmna}{boolean indicating if missing value should be accounted}

\item{metrics}{takes the value of "NES" or "ES", chooses the score to represent on the heatmap}

\item{plotly}{boolean indicating if plotly heatmaply representation or gplots heatmap should be used.}
}
\value{
Heatmap of cluster enrichment ordered by hierachical clustering
}
\description{
Heatmap of Cluster Enrichement
}
