% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection_Methods.R
\name{select_most_variable_genes}
\alias{select_most_variable_genes}
\title{Select most variable genes}
\usage{
select_most_variable_genes(X, ngenes)
}
\arguments{
\item{X}{A MCXpress class object}

\item{ngenes}{Number of genes to keep}
}
\value{
A MCXpress object with an updated Expression_Matrix
}
\description{
Select most variable genes
}

